Fixed formatting issue. Separated some math handlers. Removed isDigit method.
