+++ .idea/compiler.xml
@@ -9,5 +9,6 @@
         <module name="Discord-bot-alm" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel target="14" />
   </component>
 </project>
